refactor(application): remove unused interface bloat
Remove three categories of dead code identified through call-site analysis:

IFeatureManagementPort (entire file deleted):
- Interface had no implementation anywhere in the codebase
- 5 unused types removed: CreateForestCommand, UpdateForestCommand,
  FeatureQuery, FeatureQueryResult, IFeatureManagementPort
- RemoveFeatureCommand and FeatureOperationResult moved inline to
  RemoveFeatureUseCase (their only consumer)

IMapRepository.findByArea():
- Never called from any use case or controller
- Each implementation always returned [] (stubbed)
- Removed from domain interface and 3 infrastructure stubs:
  PostgresMapRepository, MapRepositoryAdapter, HybridMapRepository
- SpatialBounds import removed from files that no longer reference it

IMapExportPort.exportMapData():
- Convenience wrapper never called outside its own implementation
- Removed from IMapExportPort interface and MapExportService
