+++ packages/application/src/common/ports/index.ts
@@ -1,2 +1,1 @@
-export * from './IFeatureManagementPort';
 export * from './INotificationPort';
